+++ src/main/java/lotto/error/ErrorMessage.java
@@ -1,4 +1,4 @@
-package lotto.common;
+package lotto.error;
 
 public class ErrorMessage {
     public static final String MONEY_CANNOT_BE_NEGATIVE = "돈은 음수값이 될 수 없습니다.";
